Skip the whole dormant period for old tickets when calculating SLA Due date
When a customer replies to a long-dormant ticket, because Business::Hours
typically operates within a 30-day window, the business time calculation
could be incorrect and returns -1.

While it's possible to circumvent the 30-day limit by incrementally adding
time, it is practically unnecessary. In these cases, skipping the dormant
period entirely and resetting timestamp is the more efficient approach.