Align Articles autocomplete helper callback with other similar callbacks
The ModifySuggestion callback is similar in Users, Groups, and
Articles. However, the Articles callback previously passed the full
@suggestions array rather than each individual suggestion. Since
this callback is inside the loop, it runs for each value in the
collection. This would make the callback difficult to use since
it would receive the full array on each iteration.

Also, the article object was passed with the parameter "label"
rather than a more sensible "article".

Update to align with similar callbacks in Users and Groups.

Given how strange the previous parameters were, we think it's
unlikely the previous callback was being used in the wild,
so we choose not to keep the old "articles" and "label"
parameters since they would only be confusing.