Only set Resolved date when transitioning to inactive
In our RT instance we have multiple inactive states: resolved and
closed. When work is completed on a ticket it is marked as resolved, and
the Resolved date is updated accordingly. At a later stage the ticket is
moved from resolved to closed, which prevents it being re-opened.

Currently this change from resolved to closed changes the Resolved date
again. With this change, the Resolved date is only set if transitioning
from a non-inactive state to inactive (or the Resolved date isn't set).