Log unresolved ticket failures at warning log level
When calling RT::Action::SetStatus via rt-crontool, the operation fails
if the ticket has unresolved dependencies (linked tickets). This
failure is logged and previously it used the "error" level. Since
this is an informational message and not a system error, update
the log level for these "unresolved" messages to warning. All
other failures are still error.

This makes it possible for admins to better tune error notifications
and alert on errors, but not warnings based on linked tickets.