Make history options work with history search
Previously, the "Show full headers" and "Reverse history order" links only
worked on the first click because the behavior was hard-coded instead of
being a proper toggle.

Here we refactor related links and add related parameters to the history
search form. This not only fixes the above issue but also preserves the
current state during searches.