Remove the border color override and use bootstrap default
The custom color was brought forward from the RT 5 design,
but in RT 6 we are trying to use default bootstrap settings
where possible.

The color difference was noticable with the new history search
input because it was positioned directly beside the filter
button, which used the default bootstrap border color.