Restrict day evaluation to the visible calendar month
Previously, the tickets query correctly limited the tickets
to evaluate to those with dates that will show on the current
month. However, GetCalendarTickets would still process days
outside of the range if dates on the ticket started or ended
outside of the current month.

Apply the same begin/end checking for the current visible month
when processing multi-day tickets.