Fix SelfService multi-select for email inputs
This fixes a bug on the SelfService interface where when
$AllowUserAutocompleteForUnprivileged is disabled the
EmailInput fields on a ticket (Cc, Requestor etc) would be
single input fields and not multi.

They should always be multi input. Changing that resulted in
the `scomp` call to `/Helpers/Autocomplete/Principals` (bypassing
the Self Service helper) in `/Elements/EmailInput` to silently
abort due to a configuration check. Modified the helper to respect
the abort parameter when called via scomp.