Support AfterCustomFieldValue callback after code refactor
In 701adc6f48080 some per-custom-field handling code was moved
from /Elements/EditCustomFields to /Elements/EditCustomField,
including a callback that ran after each custom field. However,
we neglected to add a line with CallbackPage to allow existing
callbacks to continue to work for a period of time.

Add a callback that will still run for the previous location,
along with a deprecation message to indicate those should
be moved.