Use default bootstrap table styles and remove custom CSS
Bootstrap provides clean and consistent formats with the table
and table-bordered styles. Use those and remove custom CSS for
styles provided by those classes, specifically borders.

When using table-bordered, the table corners are not rounded,
and this can be seen in the bootstrap documentation also. The
square corners are OK for the calendar itself. It's cleaner to
stick with bootstrap defaults.