Treat all checked types as "no filter" when history filtering
If all provided types in the history filter are checked, the
history should show all normal transactions, equal to the full
history before the history filter feature was added. Previously
the provided list was being applied, which filtered out the rare
transaction types that are not provided in the "short" filter
list.