Standarize menus in titlebox headers
Previously menus contained links that had to be directly clicked,
links underlined on hover, and a dark background was shown when
clicking. This is not standard for RT menus.

Move the dropdown-item class to the correct location on the a
tag rather than the li. Apply overrides for the underline and
background on click as in other places in RT.

It also hides parent menu tooltips from dropdown elements.