Add style to allow resizing CKEditor height
The height handling changed in CKEditor 5 and as part of that,
we lost the resizing handle from CKEditor 4. There was some
extensive discussion about how to manage height in general:

https://github.com/ckeditor/ckeditor5/issues/636

And the result seems to be that they will leave the current core
behavior and leave it to CSS for customization, as noted in the
FAQ entry here:

https://ckeditor.com/docs/ckeditor5/latest/framework/how-tos.html#how-to-set-the-height-of-ckeditor-5

Add some CSS to restore the resizing handle only for vertical
resizing. The other components below the editor on the page
will move as needed.

Fixes: I#38077