Restore new as default status for cloned tickets
Status has been in the list of ticket values to propogate to
tickets cloned via the Create link in the Links section for
a very long time. However, in cb4997a4bd7 some refactoring
inadvertently stopped Status from being passed. It changed
the default to checking $DECODED_ARGS, which gets a copy of
ARGS early in the request. In Create.html, cloned values are
added to ARGS, but were not present in DECODED_ARGS, so the
create page started showing the default status for new tickets
in that queue. No one reported a bug on this change, so it
seems people like it better this way.

In RT 6, the page is built differently with htmx, so when it
is loaded, DECODED_ARGS now has many more values from the
cloned ticket. Status is one of these values and started
showing the status from the cloned ticket again.

Remove Status from the list of values to pass to restore the
previous behavior, which had been in place since RT 4.2.

Fixes: #38078