Reset inline edit row if request succeeded without data changes
To handle edge cases where a successful request results in no data changes,
we must manually clear the "refreshing" CSS class, as the inline edit row
will not be refreshed.

For example, if you try to change a ticket from "open" to "resolved" and it
fails because of some mandatory fields, then switching the status back to
"open" will not refresh the row.