Standardize on Bootstrap card body top padding
Because of legacy spacing, when RT adopted Bootstrap cards, we retained CSS
that set the top padding to zero for cards in titleboxes. To provide some
spacing, every element would then add top margin or spacing to create space
in the card body. Some of this was to allow forms and other lists to not
have special rules for the first item. This then required this extra spacing
for every new element that used a titlebox and card body.

This commit refactors CSS to use Bootstrap's default padding values where
possible. The only exception is the .card-body's padding-top where 1rem
is reduced to 0.5rem to better align with the title above (which has
padding-bottom: 0.25em) and create a more natural visual spacing.

This ensures all new titleboxes with card layouts are properly styled by
default.

For existing common titleboxes in RT, such as tables, accordions, and grid
rows, this commit adds CSS rules to ensure proper styling. For special cases
like ticket's multiple-column "Basics" layout, it applies specific Bootstrap
padding class instead.