Use saved search Rows setting when showing portlets
Saved searches rendered in dashboards long defaulted to 50 rows.
When DefaultSearchResultRowsPerPage was added, that was
referenced and a user's personal search settings. But those
were mostly intended as values for the search results page.

This commit creates a new method to return the rows setting
when rendering a portlet. It also updates the logic to reference
the rows set in the saved search, which makes sense and allows
users to manage how searches will appear in portlets from the
saved search settings.