Support user overrides for saved searches
RT has long had a page that allowed users to set custom formats,
order, and rows for searches, originally created for the searches
on the old custom home page. The home page is now a standard
dashboard, and dashboards and saved searches have been migrated
to their own tables. We can now update the search preferences
page to allow users to set preferences for any search.

Users must be Privileged and have ModifySelf. The gear icon
changes to the filled-in version to visually indicate if a
search has user preferences set.