Scroll to the top of the portlet when loading a new page
If the last page of search results has only 1 or 2 tickets,
and you are at the bottom of the search to click the page
number, when the new results load, you can end up looking
at the portlet below the one you clicked on because the
results are off the top of the screen. Scroll to the top
of the portlet to put the results in view.