Explicitly use simple on postgresql for ticket Description search strings
It ensures consistency with the FUNCTION parameter which already uses
'simple' configuration for to_tsvector(), which is also consistent with the
following index:

    CREATE INDEX Tickets7 ON Tickets USING GIN(to_tsvector('simple', Description));

This fixes a bug that can result in tickets not being returned even
if they have a searched word. Without the explicit "simple"
configuration, the default is used, typically "english" or similar
based on the system language. Simple expects full words, english can use
partial words. For example, as shown in the t/ticket/search.t test,
the search for "database" doesn't find the expected ticket because
the index stores "database" and plainto_tsquery with english
searches for "databas".