Always delay the destruction of bootstrap objects during cleanup
A JavaScript error occurred when clicking the "Edit" menu at the top right
of saved search widgets:

    Uncaught TypeError: Cannot convert undefined or null to object
        at Object.values (<anonymous>)
        at Ni._isWithActiveTrigger (bootstrap.min.js:6:48663)

This was caused by the "dispose" method being called too early, conflicting
with other active code that still required the instance. The issue can
be resolved by delaying the "dispose" method call. Previously, we had
already delayed it, but only during transitions or when it was shown, which
unfortunately wasn't sufficient.